+++ lib/danger/commands/local_helpers/local_setup.rb
@@ -28,10 +28,7 @@ module Danger
         cork.puts "Running your Dangerfile against this PR - https://#{gh.host}/#{source.repo_slug}/pull/#{source.pull_request_id}"
       end
 
-      unless verbose
-        cork.puts "Turning on --verbose"
-        dm.verbose = true
-      end
+      dm.verbose = verbose
 
       cork.puts